Overrides for Dune 2.7.1
Note however, the configure step is not present
    [
      "ocaml",
      "configure.ml",
      "--libdir",
      "#{self.lib}"
    ],

This step was meant for older opam. It looks as if we dont need it either. Adding it for Dune 2.7.1
breaks the build as this commit - https://github.com/ocaml/dune/commit/8a3d7f2f2015b71384caa07226d1a89dba9d6c25
didn't make it in the released tarball